utils/requirements.txt
mysql-connector==2.2.9
numpy==1.20.1
pandas==1.2.2
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
